+++ network/network_test.go
@@ -18,7 +18,6 @@ package network
 
 import (
 	"net/http"
-	"net/http/httptest"
 	"testing"
 )
 
@@ -41,38 +40,5 @@ func TestIsKubeletProbe(t *testing.T) {
 	req.Header.Set(KubeletProbeHeaderName, "no matter")
 	if !IsKubeletProbe(req) {
 		t.Error("kubelet probe but not counted as such")
-	}
-}
-
-func TestServeKProbe(t *testing.T) {
-	var (
-		kprobehash = "hash"
-		kprobe     = &http.Request{
-			Header: http.Header{
-				ProbeHeaderName: []string{ProbeHeaderValue},
-				HashHeaderName:  []string{kprobehash},
-			},
-		}
-		kprobeerr = &http.Request{
-			Header: http.Header{
-				ProbeHeaderName: []string{ProbeHeaderValue},
-			},
-		}
-	)
-
-	resp := httptest.NewRecorder()
-	ServeKProbe(resp, kprobe)
-	if got, want := resp.Code, http.StatusOK; got != want {
-		t.Errorf("Probe status = %d, wanted %d", got, want)
-	}
-
-	if got, want := resp.Header().Get(HashHeaderName), kprobehash; got != want {
-		t.Errorf("KProbe hash = %s, wanted %s", got, want)
-	}
-
-	resp = httptest.NewRecorder()
-	ServeKProbe(resp, kprobeerr)
-	if got, want := resp.Code, http.StatusBadRequest; got != want {
-		t.Errorf("Probe status = %d, wanted %d", got, want)
 	}
 }
